Move logging tests to test package
